# docs/swagger/swagger.yaml
basePath: /
...
definitions:
...
  github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchGroupResponse:
    description: Aggregate of matched transactions
    properties:
      confidence:
        description: Confidence score (0-100)
        example: 85
        maximum: 100
        minimum: 0
        type: integer
      confirmedAt:
        description: When the group was confirmed in RFC3339 format
        example: "2025-01-15T10:35:00Z"
        type: string
      contextId:
        description: Context ID this group belongs to
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      createdAt:
        description: Creation timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
      id:
        description: Unique identifier for the group
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      items:
        description: Matched items in this group
        items:
          $ref: '#/definitions/github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchItemResponse'
        type: array
      rejectedReason:
        description: Reason for rejection if rejected
        example: amounts do not match within tolerance
        type: string
      ruleId:
        description: Rule ID used for matching (null for manual matches)
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      runId:
        description: Run ID that created this group
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status:
        description: Current status of the group
        enum:
        - PROPOSED
        - CONFIRMED
        - REJECTED
        example: PROPOSED
        type: string
      updatedAt:
        description: Last update timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
  github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchItemResponse:
    description: Allocation of a transaction within a match group
    properties:
      allocatedAmount:
        description: Amount allocated in this match
        example: "1000.00"
        type: string
      allocatedCurrency:
        description: Currency of the allocated amount
        example: USD
        type: string
      allowPartial:
        description: Whether partial allocation is allowed
        example: false
        type: boolean
      createdAt:
        description: Creation timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
      expectedAmount:
        description: Expected amount for full match
        example: "1000.00"
        type: string
      id:
        description: Unique identifier for the item
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      matchGroupId:
        description: Match group ID this item belongs to
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      transactionId:
        description: Transaction ID being matched
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      updatedAt:
        description: Last update timestamp in RFC3339 format
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
...
  http.ErrorResponse:
    description: Standard error response returned by all API endpoints
    properties:
      code:
        description: HTTP status code
        example: 400
        type: integer
      message:
        description: Human-readable error message
        example: context name is required
        type: string
      title:
        description: Error type identifier
        example: invalid_request
        type: string
    type: object
...
  internal_matching_adapters_http.CreateManualMatchRequest:
    properties:
      notes:
        example: Manual match for Q4 reconciliation
        type: string
      transactionIds:
        example:
        - 550e8400-e29b-41d4-a716-446655440000
        items:
          type: string
        maxItems: 50
        minItems: 2
        type: array
    required:
    - transactionIds
    type: object
...
  internal_matching_adapters_http.ManualMatchResponse:
    properties:
      matchGroup:
        $ref: '#/definitions/github_com_LerianStudio_matcher_internal_matching_adapters_http_dto.MatchGroupResponse'
    type: object
...
info:
  contact:
    email: support@lerian.studio
    name: Lerian Studio Support
    url: https://discord.gg/DnhqKwkGv3
  description: |-
    Reconciliation engine for the Lerian Studio ecosystem.
    Provides automated transaction matching between Midaz ledger and external systems.
  license:
    name: Lerian Studio General License
    url: https://github.com/LerianStudio/matcher/blob/main/LICENSE.md
  termsOfService: http://swagger.io/terms/
  title: Matcher Reconciliation API
  version: v1.0.0
paths:
...
  /v1/matching/manual:
    post:
      consumes:
      - application/json
      description: Manually links multiple transactions (at least 2) into a match
        group with 100% confidence. Transactions must be UNMATCHED and belong to the
        specified context.
      operationId: createManualMatch
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Context ID
        format: uuid
        in: query
        name: contextId
        required: true
        type: string
      - description: Manual match payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_matching_adapters_http.CreateManualMatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_matching_adapters_http.ManualMatchResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Transaction not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a manual match
      tags:
      - Matching
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"